Create screenshot button (#674)
