+++ visualization/app/codeCharta/ui/screenshotButton/screenshotButton.component.html
@@ -1,0 +1,3 @@
+<md-button class="toolbar-button" ng-click="$ctrl.makeScreenshot()" title="Screenshot">
+	<i class="fa fa-camera"></i>
+</md-button>
