+++ visualization/app/codeCharta/ui/screenshotButton/screenshotButton.component.scss
@@ -1,0 +1,5 @@
+screenshot-button-component {
+	.small-action-button {
+		margin-left: 0 !important;
+	}
+}
