+++ visualization/app/codeCharta/ui/toolBar/toolBar.component.html
@@ -1,6 +1,7 @@
 <div class="tool-bar" ng-mouseup="$ctrl.onClick()">
 	<file-chooser-directive></file-chooser-directive>
 	<download-button-component ng-show="!$ctrl._viewModel.experimentalFeaturesEnabled"></download-button-component>
+	<screenshot-button-component></screenshot-button-component>
 
 	<file-panel-component ng-class="{ hidden: $ctrl._viewModel.isNodeHovered }"></file-panel-component>
 	<node-path-panel-component ng-class="{ hidden: !$ctrl._viewModel.isNodeHovered }"></node-path-panel-component>
